update error handler classes
+ Modify BaseState.dart
+ Modify BaseViewModel.dart
+ Delete FirebaseArabicErrorHandler.dart
+ Add FirebaseAuthExceptionHandler.dart
+ Delete FirebaseEnglishErrorHandler.dart
+ Add FirebaseFireStoreErrorHandler.dart
+ Add FirebaseImageDatabaseExceptionsHandler.dart
+ Add FirebaseLoginErrorHandler.dart
