+++ lib/Core/Base/BaseViewModel.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:heimdall/Core/Base/BaseNavigator.dart';
-import 'package:heimdall/Core/ErrorHandler/FirebaseArabicErrorHandler.dart';
-import 'package:heimdall/Core/ErrorHandler/FirebaseEnglishErrorHandler.dart';
+import 'package:heimdall/Core/Errors/FirebaseAuthExceptionHandler.dart';
+import 'package:heimdall/Core/Errors/FirebaseFireStoreErrorHandler.dart';
+import 'package:heimdall/Core/Errors/FirebaseImageDatabaseExceptionsHandler.dart';
+import 'package:heimdall/Core/Errors/FirebaseLoginErrorHandler.dart';
 import 'package:heimdall/Core/Providers/AppConfigProvider.dart';
 import 'package:heimdall/Core/Providers/LocalProvider.dart';
 import 'package:heimdall/Core/Providers/ThemeProvider.dart';
@@ -13,7 +15,6 @@ import 'package:heimdall/Domain/Exceptions/FirebaseUserAuthException.dart';
 import 'package:heimdall/Domain/Exceptions/InternetConnectionException.dart';
 import 'package:heimdall/Domain/Exceptions/PermissionDeniedException.dart';
 import 'package:heimdall/Domain/Exceptions/TimeOutOperationsException.dart';
-import 'package:heimdall/Domain/Exceptions/UnknownException.dart';
 import 'package:image_picker/image_picker.dart';
 
 abstract class BaseViewModel<N extends BaseNavigator> extends ChangeNotifier {
@@ -24,8 +25,11 @@ abstract class BaseViewModel<N extends BaseNavigator> extends ChangeNotifier {
   AppConfigProvider? appConfigProvider;
   AppLocalizations? local;
   Size? mediaQuery;
-  late FirebaseArabicErrorHandler firebaseArabicErrorHandler;
-  late FirebaseEnglishErrorHandler firebaseEnglishErrorHandler;
+  late FirebaseAuthExceptionHandler firebaseAuthExceptionHandler;
+  late FirebaseFireStoreErrorHandler firebaseFireStoreErrorHandler;
+  late FirebaseImageDatabaseExceptionsHandler
+      firebaseImageDatabaseExceptionsHandler;
+  late FirebaseLoginErrorHandler firebaseLoginErrorHandler;
 
   XFile? image;
 
@@ -56,22 +60,30 @@ abstract class BaseViewModel<N extends BaseNavigator> extends ChangeNotifier {
   String handleErrorMessage(Exception e) {
     if (e is FirebaseDatabaseException) {
       return localProvider!.getLocal() == "en"
-          ? firebaseEnglishErrorHandler.handleFirebaseFireStoreError(e.errorMessage)
-          : firebaseArabicErrorHandler.handleFirebaseFireStoreError(e.errorMessage);
+          ? firebaseFireStoreErrorHandler
+              .handleFirebaseFireStoreErrorEnglish(e.errorMessage)
+          : firebaseFireStoreErrorHandler
+              .handleFirebaseFireStoreErrorArabic(e.errorMessage);
     } else if (e is FirebaseImagesException) {
       return localProvider!.getLocal() == "en"
-          ? firebaseEnglishErrorHandler.handleFirebaseImageDatabaseExceptions(error: e.errorMessage,)
-          : firebaseArabicErrorHandler.handleFirebaseImageDatabaseExceptions(error: e.errorMessage,);
+          ? firebaseImageDatabaseExceptionsHandler
+              .handleFirebaseImageDatabaseExceptionsEnglish(error: e.errorMessage)
+          : firebaseImageDatabaseExceptionsHandler
+              .handleFirebaseImageDatabaseExceptionsArabic(error: e.errorMessage);
     } else if (e is FirebaseUserAuthException) {
       return localProvider!.getLocal() == "en"
-          ? firebaseEnglishErrorHandler.handleFirebaseAuthException(error: e.errorMessage,)
-          : firebaseArabicErrorHandler.handleFirebaseAuthException(error: e.errorMessage,);
-    } else if (e is FirebaseLoginException){
+          ? firebaseAuthExceptionHandler.handleFirebaseAuthExceptionEnglish(
+              error: e.errorMessage)
+          : firebaseAuthExceptionHandler.handleFirebaseAuthExceptionArabic(
+              error: e.errorMessage);
+    } else if (e is FirebaseLoginException) {
       return localProvider!.getLocal() == "en"
-          ? firebaseEnglishErrorHandler.handleLoginError(e.errorMessage,)
-          : firebaseArabicErrorHandler.handleLoginError(e.errorMessage,);
-    }else if (e is InternetConnectionException) {
-      return local!.checkYourInternetConnection ;
+          ? firebaseLoginErrorHandler
+              .handleLoginErrorEnglish(e.errorMessage)
+          : firebaseLoginErrorHandler
+              .handleLoginErrorArabic(e.errorMessage);
+    } else if (e is InternetConnectionException) {
+      return local!.checkYourInternetConnection;
     } else if (e is PermissionDeniedException) {
       return local!.weDontHavePermissions;
     } else if (e is TimeOutOperationsException) {
